requirements.txt
tqdm~=4.66.5
python-dotenv~=1.0.1
requests~=2.32.3
scholarly~=1.7.11